# db: zero mergingIter, levelIter structs before use