Previously, if a mergingIter or levelIter was reused as part of a pooled
iterAlloc, only fields explicitly set were initialized. Other fields were left
with their previous values. This caused internal iterator stats to leak between
iterators.

This commit fixes this by zeroing the entire iterAlloc struct when returning it
to the pool.