@jbowens 2022-07-26
Unfortunately, this defensive approach of zeroing the entire struct has a non-negligible affect on Iterator.Close performance:

```
name                                                             old time/op  new time/op   delta
IteratorScan/keys=100,r-amp=1,key-types=points-only-10           5.65µs ± 0%   5.90µs ± 1%  +4.52%  (p=0.008 n=5+5)
IteratorScan/keys=100,r-amp=1,key-types=points-and-ranges-10     5.84µs ± 1%   6.11µs ± 0%  +4.52%  (p=0.008 n=5+5)
IteratorScan/keys=100,r-amp=3,key-types=points-only-10           9.80µs ± 0%  10.04µs ± 0%  +2.43%  (p=0.008 n=5+5)
IteratorScan/keys=100,r-amp=3,key-types=points-and-ranges-10     10.1µs ± 1%   10.3µs ± 1%  +1.99%  (p=0.008 n=5+5)
IteratorScan/keys=100,r-amp=7,key-types=points-only-10           14.6µs ± 0%   14.8µs ± 0%  +1.36%  (p=0.008 n=5+5)
IteratorScan/keys=100,r-amp=7,key-types=points-and-ranges-10     14.8µs ± 0%   15.2µs ± 1%  +2.10%  (p=0.008 n=5+5)
IteratorScan/keys=100,r-amp=10,key-types=points-only-10          17.9µs ± 0%   18.1µs ± 0%  +1.09%  (p=0.008 n=5+5)
IteratorScan/keys=100,r-amp=10,key-types=points-and-ranges-10    18.1µs ± 0%   18.5µs ± 1%  +2.22%  (p=0.008 n=5+5)
IteratorScan/keys=1000,r-amp=1,key-types=points-only-10          43.9µs ± 0%   44.6µs ± 0%  +1.55%  (p=0.016 n=4+5)
IteratorScan/keys=1000,r-amp=1,key-types=points-and-ranges-10    46.4µs ± 0%   46.9µs ± 1%  +1.00%  (p=0.008 n=5+5)
IteratorScan/keys=1000,r-amp=3,key-types=points-only-10          77.1µs ± 1%   76.9µs ± 1%    ~     (p=0.421 n=5+5)
IteratorScan/keys=1000,r-amp=3,key-types=points-and-ranges-10    79.3µs ± 1%   78.7µs ± 0%    ~     (p=0.056 n=5+5)
IteratorScan/keys=1000,r-amp=7,key-types=points-only-10           103µs ± 0%    104µs ± 1%    ~     (p=0.421 n=5+5)
IteratorScan/keys=1000,r-amp=7,key-types=points-and-ranges-10     106µs ± 0%    106µs ± 1%  +0.70%  (p=0.008 n=5+5)
IteratorScan/keys=1000,r-amp=10,key-types=points-only-10          118µs ± 0%    118µs ± 1%    ~     (p=0.794 n=5+5)
IteratorScan/keys=1000,r-amp=10,key-types=points-and-ranges-10    120µs ± 0%    119µs ± 1%  -0.33%  (p=0.008 n=5+5)
IteratorScan/keys=10000,r-amp=1,key-types=points-only-10          417µs ± 0%    419µs ± 1%    ~     (p=0.421 n=5+5)
IteratorScan/keys=10000,r-amp=1,key-types=points-and-ranges-10    436µs ± 1%    435µs ± 1%    ~     (p=0.548 n=5+5)
IteratorScan/keys=10000,r-amp=3,key-types=points-only-10          704µs ± 1%    715µs ± 0%  +1.54%  (p=0.008 n=5+5)
IteratorScan/keys=10000,r-amp=3,key-types=points-and-ranges-10    721µs ± 1%    734µs ± 3%  +1.85%  (p=0.008 n=5+5)
IteratorScan/keys=10000,r-amp=7,key-types=points-only-10          951µs ± 1%    952µs ± 1%    ~     (p=0.690 n=5+5)
IteratorScan/keys=10000,r-amp=7,key-types=points-and-ranges-10    976µs ± 1%    981µs ± 1%    ~     (p=0.310 n=5+5)
IteratorScan/keys=10000,r-amp=10,key-types=points-only-10        1.07ms ± 2%   1.07ms ± 1%    ~     (p=0.690 n=5+5)
IteratorScan/keys=10000,r-amp=10,key-types=points-and-ranges-10  1.08ms ± 2%   1.09ms ± 1%    ~     (p=0.222 n=5+5)
```

I'm going to try just zeroing the mergingIter and the in-use levelIters when they're initialized.