@jbowens 2022-07-26
Eh, there's not much difference zeroing just the relevant `mergingIter`s and `levelIters` on init, and with zeroing during init, the overhead scales with the number of levels. I think we eat the cost of zeroing.
```
name                                                             old time/op  new time/op  delta
IteratorScan/keys=100,r-amp=1,key-types=points-only-10           5.65µs ± 0%  5.67µs ± 0%  +0.47%  (p=0.016 n=5+4)
IteratorScan/keys=100,r-amp=1,key-types=points-and-ranges-10     5.84µs ± 1%  5.90µs ± 0%  +0.91%  (p=0.032 n=5+5)
IteratorScan/keys=100,r-amp=3,key-types=points-only-10           9.80µs ± 0%  9.94µs ± 0%  +1.39%  (p=0.008 n=5+5)
IteratorScan/keys=100,r-amp=3,key-types=points-and-ranges-10     10.1µs ± 1%  10.3µs ± 0%  +1.39%  (p=0.008 n=5+5)
IteratorScan/keys=100,r-amp=7,key-types=points-only-10           14.6µs ± 0%  14.9µs ± 0%  +2.07%  (p=0.008 n=5+5)
IteratorScan/keys=100,r-amp=7,key-types=points-and-ranges-10     14.8µs ± 0%  15.3µs ± 1%  +2.89%  (p=0.008 n=5+5)
IteratorScan/keys=100,r-amp=10,key-types=points-only-10          17.9µs ± 0%  18.3µs ± 1%  +2.64%  (p=0.008 n=5+5)
IteratorScan/keys=100,r-amp=10,key-types=points-and-ranges-10    18.1µs ± 0%  18.6µs ± 0%  +2.88%  (p=0.008 n=5+5)
IteratorScan/keys=1000,r-amp=1,key-types=points-only-10          43.9µs ± 0%  44.4µs ± 1%  +0.95%  (p=0.016 n=4+5)
IteratorScan/keys=1000,r-amp=1,key-types=points-and-ranges-10    46.4µs ± 0%  46.8µs ± 1%  +0.88%  (p=0.008 n=5+5)
IteratorScan/keys=1000,r-amp=3,key-types=points-only-10          77.1µs ± 1%  77.5µs ± 0%    ~     (p=0.095 n=5+5)
IteratorScan/keys=1000,r-amp=3,key-types=points-and-ranges-10    79.3µs ± 1%  79.8µs ± 1%    ~     (p=0.056 n=5+5)
IteratorScan/keys=1000,r-amp=7,key-types=points-only-10           103µs ± 0%   104µs ± 0%  +0.74%  (p=0.008 n=5+5)
IteratorScan/keys=1000,r-amp=7,key-types=points-and-ranges-10     106µs ± 0%   107µs ± 1%  +1.01%  (p=0.008 n=5+5)
IteratorScan/keys=1000,r-amp=10,key-types=points-only-10          118µs ± 0%   117µs ± 0%  -0.31%  (p=0.008 n=5+5)
IteratorScan/keys=1000,r-amp=10,key-types=points-and-ranges-10    120µs ± 0%   120µs ± 0%    ~     (p=0.190 n=5+5)
IteratorScan/keys=10000,r-amp=1,key-types=points-only-10          417µs ± 0%   417µs ± 1%    ~     (p=0.310 n=5+5)
IteratorScan/keys=10000,r-amp=1,key-types=points-and-ranges-10    436µs ± 1%   440µs ± 0%  +0.93%  (p=0.032 n=5+5)
IteratorScan/keys=10000,r-amp=3,key-types=points-only-10          704µs ± 1%   705µs ± 2%    ~     (p=0.841 n=5+5)
IteratorScan/keys=10000,r-amp=3,key-types=points-and-ranges-10    721µs ± 1%   737µs ± 0%  +2.32%  (p=0.008 n=5+5)
IteratorScan/keys=10000,r-amp=7,key-types=points-only-10          951µs ± 1%   958µs ± 1%    ~     (p=0.095 n=5+5)
IteratorScan/keys=10000,r-amp=7,key-types=points-and-ranges-10    976µs ± 1%   973µs ± 1%    ~     (p=0.421 n=5+5)
IteratorScan/keys=10000,r-amp=10,key-types=points-only-10        1.07ms ± 2%  1.07ms ± 1%    ~     (p=0.690 n=5+5)
IteratorScan/keys=10000,r-amp=10,key-types=points-and-ranges-10  1.08ms ± 2%  1.09ms ± 0%    ~     (p=0.151 n=5+5)
```