@jbowens 2022-07-26
Going to backport this to 22.1, since iterator stats were exposed in 22.1.